+++ src/root.rs
@@ -158,7 +158,7 @@ pub mod bench {
     fn bench_root<const BITS: usize, const LIMBS: usize>(criterion: &mut Criterion, degree: usize) {
         let input = Uint::<BITS, LIMBS>::arbitrary();
         let mut runner = TestRunner::deterministic();
-        criterion.bench_function(&format!("root/{}/{}", degree, BITS), move |bencher| {
+        criterion.bench_function(&format!("root/{degree}/{BITS}"), move |bencher| {
             bencher.iter_batched(
                 || input.new_tree(&mut runner).unwrap().current(),
                 |value| black_box(black_box(value).root(black_box(degree))),